Restructure files to processor folder
